Portfolio Cleanup
I cleaned up the portfolio area by getting rid of the "photo-container" div, which was unnecessary.  I moved the properties into the "photo" tag to have the same functionality.  I also cleaned up the CSS file to have the overlay come in as visible/hidden rather than the opacity change on hover and reorganized it.